minor refactors

+++ src/main/java/com/startlet/starlet_academy/repositorys/ParentRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ParentRepository extends JpaRepository<Parent, Integer> {
+public interface ParentRepository extends JpaRepository<Parent, Long> {
     @Query(value = "SELECT * FROM Payment p WHERE p.fee_id = :feeId",nativeQuery = true)
     List<Payment> findByFeeFeeId(int feeId);
